server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-oscar-bundles"
version = "2.3.0"
description = "An extension on top of django-oscar providing product bundle functionality."
readme = "README.rst"
license = { text = "ISC" }
dependencies = [
    "Django (>=4.2)",
    "django-oscar (>=4.0,<4.1)",
    "django-oscar-api (==3.3.0)",
    "phonenumbers (>=9.0.3,<10)",
    "django-stubs-ext (>=5.1.3,<6)",
]
requires-python = ">=3.12"

[[project.authors]]
name = "thelab"
email = "thelabdev@thelab.co"

[project.urls]
Homepage = "https://gitlab.com/thelabnyc/django-oscar/django-oscar-bundles/"
Repository = "https://gitlab.com/thelabnyc/django-oscar/django-oscar-bundles/"

[dependency-groups]
dev = [
    "coverage (==7.9.1)",
    "flake8 (==7.3.0)",
    "ipdb (==0.13.13)",
    "psycopg2-binary (==2.9.10)",
    "sorl-thumbnail (==12.11.0)",
    "tox (==4.27.0)",
    "unittest-xml-reporting (==3.2.0)",
    "mypy (==1.16.1)",
    "django-stubs (==5.2.1)",
    "djangorestframework-stubs (==3.16.0)",
    "tox-uv>=1.26.1",
]

[tool.hatch.build.targets.wheel]
packages = ["oscarbundles"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

# Strict mode, see mypy --help
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
# disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
show_error_codes = true
# Not turned on by strict
strict_equality = true

[[tool.mypy.overrides]]
module = "oscar.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.tests.*"
ignore_errors = true


[tool.django-stubs]
django_settings_module = "sandbox.settings"

[tool.isort]
profile = "black"
from_first = true

[tool.coverage.run]
branch = true
source_pkgs = ["oscarbundles"]

[tool.coverage.report]
show_missing = true
ignore_errors = true
